LPD-42546 Update tests to use new navigation, simplify setups, merge non-breaking tests, group tests on "Customized" sample into one file
